backend/requirements.txt
django==4.1.1
djangorestframework==3.13.1
drf_yasg==1.21.3